refactor(ors_control_app): move ensureBackload init helper to lib/init.ts
Pre-install cleanup. Move the 280-line ensureBackloadAndAssetVelocityObjects
helper out of server/index.ts into a new server/lib/init.ts module:
- Single export consumed by index.ts (boot-init block) and routes/fleet.ts
  (next commit, for /api/backload/seed).
- Imports currentRegionScalar from lib/region.ts and log from diagnostics.

server/index.ts shrinks from 525 -> 225 lines and is now closer to the
plan target of bootstrap + verifySessionUtc + app.listen + router mounts.

No behaviour change. tsc passes cleanly.
